+++ spec/rubocop/cop/lint/string_conversion_in_interpolation_spec.rb
@@ -7,6 +7,12 @@ RSpec.describe RuboCop::Cop::Lint::StringConversionInInterpolation do
     expect_offense(<<-'RUBY'.strip_indent)
       "this is the #{result.to_s}"
                             ^^^^ Redundant use of `Object#to_s` in interpolation.
+      /regexp #{result.to_s}/
+                       ^^^^ Redundant use of `Object#to_s` in interpolation.
+      :"symbol #{result.to_s}"
+                        ^^^^ Redundant use of `Object#to_s` in interpolation.
+      `backticks #{result.to_s}`
+                          ^^^^ Redundant use of `Object#to_s` in interpolation.
     RUBY
   end